+++ src/api/auth/postSignUp.ts
@@ -1,28 +1,6 @@
 import basicAxios from "@/api/basicAxios";
-import { SignUpData } from "../types/kakao";
-import { getKakaoAccessToken, removeKakaoAccessToken, setKakaoAccessToken } from "@/utils/storage";
-
-export const postKakaoLogin = async (kakaoAccessToken:string) => {
-  const endpoint = '/api/auth/kakao-login';
-
-  try {
-    const response = await basicAxios.post(
-      endpoint,
-      {},
-      {
-        headers: {
-          Authorization: `Bearer ${kakaoAccessToken}`,
-        },
-      }
-    );
-    console.log('카카오 로그인 성공:', response);
-    return response.data;
-  } catch (error) {
-    console.error('카카오 로그인 실패:', error);
-    setKakaoAccessToken(kakaoAccessToken);
-    throw error;
-  }
-};
+import { getKakaoAccessToken, removeKakaoAccessToken } from "@/utils/storage";
+import { SignUpData } from "./types/kakao";
 
 export const postSignUp = async ({
   serviceTerm,
