--- src/features/Badge/constants/badge.ts
@@ -1,85 +1,0 @@
-import { Badge } from '../types/Badge';
-
-export const BADGES: Badge[] = [
-  {
-    id: 0,
-    name: '첫 꿀 받기',
-    goal: '받은 꿀 1회',
-    image: '/assets/images/badge/badge-get-1.svg',
-    isObtain: true,
-    date: '2024-11-16',
-  },
-  {
-    id: 1,
-    name: '꿀 좀 받아본 사람',
-    goal: '받은 꿀 7회',
-    image: '/assets/images/badge/badge-get-1.svg',
-    isObtain: true,
-    date: '2024-11-16',
-  },
-  {
-    id: 2,
-    name: '꿀부자',
-    goal: '받은 꿀 50회',
-  },
-  {
-    id: 3,
-    name: '첫 꿀 보내기',
-    goal: '보낸 꿀 1회',
-  },
-  {
-    id: 4,
-    name: '달달한 당신',
-    goal: '보낸 꿀 7회',
-  },
-  {
-    id: 5,
-    name: '꿀플루언서',
-    goal: '보낸 꿀 50회',
-  },
-  {
-    id: 6,
-    name: '알찬 마음',
-    goal: '꿀도장 1판 달성',
-  },
-  {
-    id: 7,
-    name: '꽉 찬 도장판',
-    goal: '꿀도장 3판 달성',
-  },
-  {
-    id: 8,
-    name: '꿀만장자',
-    goal: '꿀도장 10판 달성',
-  },
-  {
-    id: 9,
-    name: '친해지는 중',
-    goal: '출석 누적 5회',
-  },
-  {
-    id: 10,
-    name: '다정한 사이',
-    goal: '출석 누적 30회',
-  },
-  {
-    id: 11,
-    name: '온 세상이 꿀',
-    goal: '출석 누적 100회',
-  },
-  {
-    id: 12,
-    name: '???',
-    goal: '???',
-  },
-  {
-    id: 13,
-    name: '???',
-    goal: '???',
-  },
-  {
-    id: 14,
-    name: '???',
-    goal: '???',
-  },
-];
